fixed s3-aws storage driver
