+++ engine/extensions/storage-drivers/implementations/s3-aws/src/main/resources/META-INF/services/com.dell.spt.storage.driver.base.StorageDriverFactory
@@ -1,0 +1,1 @@
+com.dell.spt.storage.driver.coop.aws.s3.S3AwsStorageDriverFactory
