+++ engine/extensions/storage-drivers/implementations/s3-aws/src/main/resources/config-schema-storage-s3-aws.yaml
@@ -1,0 +1,14 @@
+---
+# WARNING: Do not edit this file manually, it will be replaced with factory default content on every Spt run.
+storage:
+  auth:
+    version: int
+  checksum:
+    enabled: boolean
+    algorithm: string
+  object:
+    fsAccess: boolean
+    tagging:
+      enabled: boolean
+      tags: map
+    versioning: boolean
