+++ engine/extensions/storage-drivers/implementations/s3-aws/src/main/resources/config/defaults-storage-s3-aws.yaml
@@ -1,0 +1,15 @@
+---
+# WARNING: Do not edit this file manually, it will be replaced with factory default content on every Spt run. Use
+# the command line arguments either load step configurations in the scenario files instead.
+storage:
+  auth:
+    version: 2
+  checksum:
+    enabled: false
+    algorithm: md5
+  object:
+    fsAccess: false
+    tagging:
+      enabled: false
+      tags: {}
+    versioning: false
